Added the fancy list that Pete designed as an option within the pic and text block. You can select the bullet points instead of the wysiwyg editor. The layout for the bullet points is in content-fancyBullets.php and style.css. The fancy bullets are currently included as an option in content-picAndText.php.
I also did a bunch of CSS fixes. I fixed a bug that was causing the menu nav to be hidden on screens below a certain size (it was leftover CSS from 2017 causing it). I also made the menu full screen and optomised it so the items would be indented with margins of about 40% that shrink back to the default of 5% on smaller screens. I also removed the decoridritve divs from the bottom of the menu (jen's request)

I removed the border from the mission div at jen's request.